backend/requirements.txt
fastapi==0.115.5
pydantic[email]==2.13.1
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.28.0
apscheduler==3.10.4
